+++ src/main/groovy/com/cedarsoftware/controller/NCubeController.groovy
@@ -190,19 +190,7 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
     Object[] search(ApplicationID appId, String cubeNamePattern = null, String content = null, Map options = [(SEARCH_ACTIVE_RECORDS_ONLY):true])
     {
         appId = addTenant(appId)
-        if (cubeNamePattern != null)
-        {
-            cubeNamePattern = cubeNamePattern.trim()
-        }
         List<NCubeInfoDto> cubeInfos = mutableClient.search(appId, cubeNamePattern, content, options)
-
-        Collections.sort(cubeInfos, new Comparator<NCubeInfoDto>() {
-            int compare(NCubeInfoDto info1, NCubeInfoDto info2)
-            {
-                return info1.name.compareToIgnoreCase(info2.name)
-            }
-        })
-
         return cubeInfos as Object[]
     }