+++ src/main/groovy/com/cedarsoftware/ncube/NCubeRuntime.groovy
@@ -187,10 +187,23 @@ class NCubeRuntime implements NCubeMutableClient, NCubeRuntimeClient, NCubeTestC
      */
     List<NCubeInfoDto> search(ApplicationID appId, String cubeNamePattern, String content, Map options)
     {
-        Object[] searchResults = bean.call(beanName, 'search', [appId, cubeNamePattern, content, options]) as Object[]
-        List<NCubeInfoDto> dtos = []
-        searchResults.each { NCubeInfoDto dto -> dtos.add(dto) }
-        return dtos
+        if (cubeNamePattern != null)
+        {
+            cubeNamePattern = cubeNamePattern.trim()
+        }
+        if (options && options.containsKey(SEARCH_CLOSURE))
+        {
+            options.remove(SEARCH_CLOSURE)
+            options.remove(SEARCH_OUTPUT)
+        }
+        List<NCubeInfoDto> cubeInfos = (List<NCubeInfoDto>) bean.call(beanName, 'search', [appId, cubeNamePattern, content, options])
+        Collections.sort(cubeInfos, new Comparator<NCubeInfoDto>() {
+            int compare(NCubeInfoDto info1, NCubeInfoDto info2)
+            {
+                return info1.name.compareToIgnoreCase(info2.name)
+            }
+        })
+        return cubeInfos
     }
 
     /**